removed templates from engine logic and added to templates file
